First Commit
This is my first commit for project 1, I should had built the
repository sooner and saved/pushed along the way but this is a very
nice starting point nonetheless. I added in what I believe should be
all the requirements for the project 1 assignment. I will probably
double check, cleanup and add more comments tomorrow before turning in
the assignment.