Project 2 - Save 2
Added code to main.js javascript file, can't seem to link the
javascript to the additem.html page. It's probably something simple
that I am overlooking here. I have added the <script> tag at the bottom
of the additem.html page right before the closing </body> tag like in
the video but it doesn't seem to be linking correctly. Again, it's
something simple I just know it. I'll have to overlook more this
evening; as it's hard to tell what is working and what needs fixed
since I can't even test or debug the js.