Actual Final Save
I could never ever get over the debugging hump and I don't know why,
I've fixed a lot of issues but more came and so on, it was just a never
ending battle and because of my scheduling I couldn't email instructor
til late for help on Weds. night, maybe around 7-8, I don't know but
couldn't get a response back, which is my own fault because of current
scheduling.

I tried to use JSLint but that only got me so far. I have failed so
horribly here because of the issue of a click listener event. I should
just give up I guess and learn development as a hobby? I don't know.